\name{lola_kn}
\alias{lola_kn}
\docType{data}
\title{spatial locations of 1535 weather stations}
\description{
   The dataset contains spatial locations of 1535 weather stations 
   for measuring solar irradiance across the continental United States.
}
\usage{data(lola_kn)}
\format{A data frame containing 1535 observations and 2 variables}
\source{\url{https://onlinelibrary.wiley.com/doi/abs/10.1002/sam.11414}}
\references{
  F. Sun, R.B. Gramacy, B. Haaland, S.Y. Lu, and Y. Hwang (2019)
  \emph{Synthesizing Simulation and Field Data of Solar Irradiance}, Statistical Analysis and Data Mining, 12(4), 311-324;
  preprint on arXiv:1806.05131.
}
\keyword{datasets}